requirements.txt
mesa==3.0.5
neo4j==5.24.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
networkx==3.3
openai==1.37.0
langchain==0.2.11
langchain-openai==0.1.17
chromadb==0.5.5
fastapi==0.111.1
uvicorn==0.30.3
python-dotenv==1.0.1
pydantic==2.8.2
matplotlib==3.9.1
plotly==5.23.0
community==1.0.0b1
sentence-transformers==3.0.1